feat(storage): handle comparisons between record and different object
We return a `NotImplemented` singleton value for built-in comparison
methods to tell Python to check the reverse comparison if the compared
object is of a different class. Without this, it would return True or
False as long as no error is raised, but that could lead to
inconsistencies (comparing things that should not be) without being
warned.

We raise a `TypeError` for the `is_duplicate` method, because this
method should never be called on objects of different classes, and,
contrary to built-in comparison methods, there is no reverse operation.